mv some dataclasses around
