+++ scripts/pipeline_example.py
@@ -47,40 +47,11 @@ import re
 from dataclasses import dataclass
 from typing import Any, Callable, List
 
-# #####################################################################################
-# #######################START LANGUAGE TYPE###########################################
+from cltkv1.languages.glottolog import LANGUAGES
+from cltkv1.utils.data_types import Language, Word
+from cltkv1.utils.operations import Operation
+from cltkv1.nlp import NLP
 
-
-@dataclass
-class Language:
-    name: str
-    glottocode: str
-    latitude: float
-    longitude: float
-    dates: List[float]
-    family_id: str
-    parent_id: str
-    level: str
-    iso639P3code: str
-    type: str
-
-
-LATIN = Language(
-    name="Latin",
-    glottocode="lati1261",
-    latitude=60.2,
-    longitude=50.5,
-    dates=[200, 400],
-    family_id="indo1319",
-    parent_id="impe1234",
-    level="language",
-    iso639P3code="lat",
-    type="a",
-)
-
-
-# #########################END LANGUAGE TYPE###########################################
-# #####################################################################################
 
 
 # #####################################################################################
@@ -103,22 +74,6 @@ def dummy_get_token_indices(text: str) -> List[List[int]]:
     return indices_words
 
 
-@dataclass
-class Operation:
-    """For each type of NLP operation there needs to be a definition.
-    It includes the type of data it expects (``str``, ``List[str]``,
-    ``Word``, etc.) and what field withing ``Word`` it will populate.
-    This base class is intended to be inherited by NLP operation
-    types (e.g., ``TokenizationOperation`` or ``DependencyOperation``).
-    """
-
-    name: str
-    description: str
-    input: Any
-    output: Any
-    algorithm: Callable
-    type: str
-
 
 @dataclass
 class TokenizationOperation(Operation):
@@ -139,33 +94,12 @@ class LatinTokenizationOperation(TokenizationOperation):
     input = str
     output = List[List[int]]  # e.g., [[0, 4], [6, 11], ...]
     algorithm = dummy_get_token_indices
-    language = LATIN
+    language = LANGUAGES["lat"]
 
 
 # #######################END OPERATION TYPE############################################
 # #####################################################################################
 
-
-# #####################################################################################
-# #######################START WORD TYPE###############################################
-
-
-@dataclass
-class Word:
-    """Contains attributes of each processed word in a list of tokens. To be used most often in the ``Doc.tokens``
-    dataclass. """
-
-    index_char_start: int = None
-    index_char_stop: int = None
-    index_token: int = None
-    index_sentence: int = None
-    string: str = None
-    pos: str = None
-    scansion: str = None
-
-
-# #####################################################################################
-# #######################END WORD TYPE#################################################
 
 
 # #####################################################################################
@@ -185,108 +119,17 @@ class Pipeline:
 class LatinPipeline(Pipeline):
     # sentence_splitter = LatinSplitter().dummy_get_indices
     word_tokenizer = LatinTokenizationOperation
-    language = LATIN
+    language = LANGUAGES["lat"]
 
 
 # #######################END PIPELINE TYPE#############################################
 # #####################################################################################
 
 
-# #####################################################################################
-# #######################START DOC TYPE################################################
-
-
-@dataclass
-class Doc:
-    """The object returned to the user from the ``NLP()`` class. Contains overall attributes of submitted texts,
-    plus most importantly the processed tokenized text ``tokens``, being a list of ``Word`` types.. """
-
-    indices_sentences: List[List[int]] = None
-    indices_tokens: List[List[int]] = None
-    language: str = None
-    tokens: List[Word] = None
-    pipeline: Any = None
-    raw: str = None
-    ner: List[List[int]] = None
-
-
-# #######################END DOC TYPE##################################################
-# #####################################################################################
-
-
-# #####################################################################################
-# #######################START NLP CLASS###############################################
-
-
-class NLP:
-    """A dummy example of the primary entry-point class."""
-
-    def __init__(self, language: str) -> None:
-        self.language = language
-        if self.language == "latin":
-            self.pipeline = LatinPipeline
-        else:
-            raise NotImplementedError(
-                f"Pipeline not available for language '{self.language}'."
-            )
-
-    def run_pipeline(self, text: str) -> Doc:
-        """Take a raw unprocessed text string, then return a ``Doc`` object
-        containing all available processed information.
-        """
-        # Get token indices
-        token_indices = self.pipeline.word_tokenizer.algorithm(text=text)
-
-        # Populate a ``Word`` object for each token in the submitted text
-        all_word_tokens = list()
-        for token_count, token_index in enumerate(token_indices):
-            token_start = token_index[0]
-            token_end = token_index[1]
-            token_str = text[token_start:token_end]
-
-            # index_char_start: int = None
-            # index_char_stop: int = None
-            # index_token: int = None
-            # index_sentence: int = None
-            # string: str = None
-            # pos: str = None
-            # scansion: str = None
-
-            word = Word(
-                index_char_start=token_start,
-                index_char_stop=token_end,
-                index_token=token_count,
-                string=token_str,
-            )
-            all_word_tokens.append(word)
-
-        doc = Doc(
-            indices_tokens=token_indices,
-            language=self.language,
-            pipeline=self.pipeline,
-            tokens=all_word_tokens,
-            raw=text,
-        )
-
-        return doc
-
-
-# #######################END NLP CLASS#################################################
-# #####################################################################################
-
-
 if __name__ == "__main__":
-    tacitus_germanica = (
-        "Germania omnis a Gallis Raetisque et Pannoniis Rheno et Danuvio fluminibus, a Sarmatis "
-        "Dacisque mutuo metu aut montibus separatur: cetera Oceanus ambit, latos sinus et insularum "
-        "inmensa spatia complectens, nuper cognitis quibusdam gentibus ac regibus, quos bellum "
-        "aperuit. Rhenus, Raeticarum Alpium inaccesso ac praecipiti vertice ortus, modico flexu in "
-        "occidentem versus septentrionali Oceano miscetur. Danuvius molli et clementer edito montis "
-        "Abnobae iugo effusus pluris populos adit, donec in Ponticum mare sex meatibus erumpat: "
-        "septimum os paludibus hauritur. "
-    )
+    from cltkv1.utils.example_texts import LATIN
     cltk_nlp = NLP(language="latin")
-    doc_germanica = cltk_nlp.run_pipeline(tacitus_germanica)
+    doc_germanica = cltk_nlp.run_pipeline(LATIN)
 
     print("")
     print("``Doc``:", doc_germanica)
